Add interface for RoutingTable for testing
