+++ src/StaticRouter.cpp
@@ -4,13 +4,13 @@
 #include <cstring>
 
 #include "protocol.h"
-#include "IPacketSender.h"
 #include "utils.h"
 
-StaticRouter::StaticRouter(std::unique_ptr<IArpCache> arpCache, std::shared_ptr<RoutingTable> routingTable, std::shared_ptr<IPacketSender> packetSender)
-: routingTable(routingTable)
-, packetSender(packetSender)
-, arpCache(std::move(arpCache))
+StaticRouter::StaticRouter(std::unique_ptr<IArpCache> arpCache, std::shared_ptr<IRoutingTable> routingTable,
+                           std::shared_ptr<IPacketSender> packetSender)
+    : routingTable(routingTable)
+      , packetSender(packetSender)
+      , arpCache(std::move(arpCache))
 {
 }
 
@@ -18,11 +18,11 @@ void StaticRouter::handlePacket(std::vector<uint8_t> packet, std::string iface)
 {
     std::unique_lock lock(mutex);
 
-    if (packet.size() < sizeof(sr_ethernet_hdr_t)) {
+    if (packet.size() < sizeof(sr_ethernet_hdr_t))
+    {
         spdlog::error("Packet is too small to contain an Ethernet header.");
         return;
     }
 
     // TODO: Your code below
-
 }
